refactor(TabManager): Update import package, updated to use injected configuration file, ignore scheduler if option is disabled. Added to injector binder.
