+++ plugin/src/main/java/com/xism4/sternalboard/listener/module/ListenerModule.java
@@ -1,6 +1,7 @@
 package com.xism4.sternalboard.listener.module;
 
 import com.xism4.sternalboard.listener.ScoreboardListener;
+import com.xism4.sternalboard.listener.TabListener;
 import org.bukkit.event.Listener;
 import team.unnamed.inject.AbstractModule;
 
@@ -8,6 +9,10 @@ public class ListenerModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        super.multibind(Listener.class).asSet().to(ScoreboardListener.class).singleton();
+        super.multibind(Listener.class)
+                .asSet()
+                .to(TabListener.class)
+                .to(ScoreboardListener.class)
+                .singleton();
     }
 }
